+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added a `with_image()` helper to mocked HTTP responses.
+- Added a `is_blob()` and `is_file()` helper to `Mantle\Http_Client\Response`.
+
 ### Fixed
 
 - Fixed an issue with taxonomy registration not returning an array.